move e2e test utils under @angular/cdk/testing
