+++ src/cdk/testing/e2e/BUILD.bazel
@@ -1,0 +1,11 @@
+package(default_visibility=["//visibility:public"])
+
+load("//tools:defaults.bzl", "ng_e2e_test_library")
+
+exports_files(["tsconfig-e2e.json"])
+
+ng_e2e_test_library(
+  name = "e2e",
+  srcs = glob(["**/*.ts"]),
+  module_name = "@angular/cdk/testing/e2e"
+)
