+++ src/cdk/tsconfig-tests.json
@@ -34,6 +34,9 @@
     "**/*.spec.ts"
   ],
   "exclude": [
-    "**/schematics/**/*.ts"
+    "**/schematics/**/*.ts",
+    // Exclude end-to-end tests and utilities
+    "**/*.e2e.spec.ts",
+    "./testing/e2e/**"
   ]
 }
