+++ src/e2e-app/BUILD.bazel
@@ -5,17 +5,7 @@ load("@npm_bazel_typescript//:defs.bzl", "ts_devserver", "ts_library")
 load("//tools:defaults.bzl", "ng_module")
 load("//:packages.bzl", "ANGULAR_LIBRARY_UMDS")
 
-exports_files(["protractor.conf.js", "start-devserver.js", "tsconfig-e2e.json"])
-
-ts_library(
-  name = "e2e_utils_lib",
-  srcs = glob(["test-util/**/*.ts"]),
-  tsconfig = ":tsconfig-e2e.json",
-  deps = [
-    "@npm//@types/jasmine",
-    "@npm//protractor"
-  ]
-)
+exports_files(["protractor.conf.js", "start-devserver.js"])
 
 ng_module(
   name = "e2e-app",
