--- src/e2e-app/tsconfig-e2e.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["es2015"],
-    "types": [
-      "jasmine"
-    ]
-  }
-}
