+++ src/material/tsconfig-tests.json
@@ -29,6 +29,7 @@
     "index.ts"
   ],
   "exclude": [
-    "**/schematics/**/*.ts"
+    "**/schematics/**/*.ts",
+    "**/*.e2e.spec.ts"
   ]
 }
